:building_construction: Follow DRY
