+++ packages/client/pages/overview.vue
@@ -14,6 +14,7 @@ import SlideContainer from '../internals/SlideContainer.vue'
 import SlideWrapper from '../internals/SlideWrapper.vue'
 import { isColorSchemaConfigured, isDark, toggleDark } from '../logic/dark'
 import { getSlidePath } from '../logic/slides'
+import { wordCount } from '../logic/utils'
 
 const cardWidth = 450
 
@@ -46,23 +47,6 @@ function toggleRoute(route: SlideRoute) {
     activeSlide.value = undefined
   else
     activeSlide.value = route
-}
-
-function wordCount(str: string) {
-  const pattern = /[\w`'\-\u0392-\u03C9\u00C0-\u00FF\u0600-\u06FF\u0400-\u04FF]+|[\u4E00-\u9FFF\u3400-\u4DBF\uF900-\uFAFF\u3040-\u309F\uAC00-\uD7AF]+/g
-  const m = str.match(pattern)
-  let count = 0
-  if (!m)
-    return 0
-  for (let i = 0; i < m.length; i++) {
-    if (m[i].charCodeAt(0) >= 0x4E00) {
-      count += m[i].length
-    }
-    else {
-      count += 1
-    }
-  }
-  return count
 }
 
 function isElementInViewport(el: HTMLElement) {
