+++ packages/client/pages/overview/print.vue
@@ -10,6 +10,7 @@ import { configs } from '../../env'
 import NoteDisplay from '../../internals/NoteDisplay.vue'
 import SlideContainer from '../../internals/SlideContainer.vue'
 import SlideWrapper from '../../internals/SlideWrapper.vue'
+import { wordCount } from '../../logic/utils'
 
 const cardWidth = 400
 
@@ -42,21 +43,6 @@ function getClicksContext(route: SlideRoute) {
   if (!clicksContextMap.has(route))
     clicksContextMap.set(route, createFixedClicks(route, CLICKS_MAX))
   return clicksContextMap.get(route)!
-}
-
-function wordCount(str: string) {
-  const pattern = /[\w`'\-\u0392-\u03C9\u00C0-\u00FF\u0600-\u06FF\u0400-\u04FF]+|[\u4E00-\u9FFF\u3400-\u4DBF\uF900-\uFAFF\u3040-\u309F\uAC00-\uD7AF]+/g
-  const m = str.match(pattern)
-  let count = 0
-  if (!m)
-    return 0
-  for (let i = 0; i < m.length; i++) {
-    if (m[i].charCodeAt(0) >= 0x4E00)
-      count += m[i].length
-    else
-      count += 1
-  }
-  return count
 }
 
 const wordCounts = computed(() => slides.value.map(route => wordCount(route.meta?.slide?.note || '')))
